+++ client/index.html
@@ -1,4 +1,4 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
@@ -8,6 +8,7 @@
   </head>
   <body>
     <div id="root"></div>
+    <script src="https://unpkg.com/react-scan/dist/auto.global.js"></script>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
