+++ client/src/providers/DataProvider.tsx
@@ -1,0 +1,29 @@
+import { useState, useEffect } from "react";
+import { DataContext } from "../contexts/DataContext";
+import type { BankProduct } from "../contexts/DataContext";
+
+export const DataProvider = ({ children }: { children: React.ReactNode }) => {
+  const [data, setData] = useState<BankProduct[]>([]);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const res = await fetch("http://localhost:3333/");
+        const json = await res.json();
+        setData(json);
+      } catch (err) {
+        console.error("데이터 불러오기 실패:", err);
+      }
+    };
+
+    fetchData();
+  }, []);
+
+  return (
+    <>
+      {data.length > 0 && (
+        <DataContext value={{ data, setData }}>{children}</DataContext>
+      )}
+    </>
+  );
+};
